+++ src/commands/open-settings.js
@@ -2,7 +2,10 @@ const vscode = require("vscode");
 
 function openSettings() {
   return vscode.commands.registerCommand("gitmob.openSettings", function () {
-    vscode.commands.executeCommand("workbench.action.openSettings", "gitmob");
+    vscode.commands.executeCommand(
+      "workbench.action.openSettings",
+      "@ext:richardkotze.git-mob"
+    );
   });
 }